requirements.txt
apache-airflow-providers-postgres>=6.0.0
apache-airflow-providers-mysql>=6.0.0
apache-airflow-providers-snowflake>=6.0.0
apache-airflow-providers-common-sql>=1.16.0
psycopg2-binary>=2.9.9
pymysql>=1.1.1
pandas>=2.2.0
duckdb>=1.1.0
duckdb-engine>=0.13.0
snowflake-connector-python[pandas]>=3.12.0
streamlit>=1.39.0
plotly>=5.24.0
requests>=2.32.0
python-dotenv>=1.0.1
